Used this version to make v0.2.
Removed the '--no-install-recommends' from the apt-get install commands.
Added the git package.
Added the plug.vim file. This installs the vim-plug "package manager".
Added a 'vim' user and the necessary 'chown' command.
Added a command that installs the vim plugins for the vim user.
Enhanced the .vimrc file to load the following plugins:
   CtrlP
   Syntastic
   wombat256.vim
   vim-airline
Made changes to .vimrc to support the new plugins.